backend/agents/execution_agent/exec_agent_requirements.txt
# Core Dependencies
pywinauto>=0.6.8
pyautogui>=0.9.53
Pillow>=9.0.0

# Web Automation
selenium>=4.0.0
playwright>=1.30.0

# OCR
pytesseract>=0.3.10

# Optional but recommended
python-dateutil>=2.8.2
